// src/generated/data-contracts.ts
export interface GraphStatsData {
  /** @format uint64 */
  cyberlinks?: string;
  /** @format uint64 */
  particles?: string;
}

export interface IsAnyLinkExistData {
  /** @format boolean */
  exist?: boolean;
}

export interface IsLinkExistData {
  /** @format boolean */
  exist?: boolean;
}

export interface KarmaData {
  /** @format uint64 */
  karma?: string;
}

export interface NegentropyData {
  /** @format uint64 */
  negentropy?: string;
}

export interface ParticleNegentropyData {
  /** @format uint64 */
  entropy?: string;
}

export interface BacklinksData {
  result?: {
    particle?: string;
    /** @format uint64 */
    rank?: string;
  }[];
  pagination?: {
    /** @format int64 */
    total?: number;
  };
}

export interface RankParamsData {
  params?: {
    /** @format int64 */
    calculation_period?: string;
    damping_factor?: string;
    tolerance?: string;
  };
}

export interface RankData {
  /** @format uint64 */
  rank?: string;
}

export interface SearchData {
  result?: {
    particle?: string;
    /** @format uint64 */
    rank?: string;
  }[];
  pagination?: {
    /** @format int64 */
    total?: number;
  };
}

export interface TopData {
  result?: {
    particle?: string;
    /** @format uint64 */
    rank?: string;
  }[];
  pagination?: {
    /** @format int64 */
    total?: number;
  };
}

export interface LoadData {
  /** DecProto defines a Protobuf wrapper around a Dec object. */
  load?: {
    dec?: string;
  };
}

export interface NeuronBandwidthData {
  neuron_bandwidth?: {
    neuron?: string;
    /** @format uint64 */
    remained_value?: string;
    /** @format uint64 */
    last_updated_block?: string;
    /** @format uint64 */
    max_value?: string;
  };
}

export interface BandwidthParamsData {
  params?: {
    /** @format uint64 */
    recovery_period?: string;
    /** @format uint64 */
    adjust_price_period?: string;
    base_price?: string;
    base_load?: string;
    /** @format uint64 */
    max_block_bandwidth?: string;
  };
}

export interface PriceData {
  /** DecProto defines a Protobuf wrapper around a Dec object. */
  price?: {
    dec?: string;
  };
}

export interface TotalBandwidthData {
  /** @format uint64 */
  total_bandwidth?: string;
}

export interface DestinationRoutedEnergyData {
  value?: {
    denom?: string;
    amount?: string;
  }[];
}

export interface DestinationRoutesData {
  routes?: {
    source?: string;
    destination?: string;
    name?: string;
    value?: {
      denom?: string;
      amount?: string;
    }[];
  }[];
  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: {
    /**
     * next_key is the key to be passed to PageRequest.key to
     * query the next page most efficiently
     * @format byte
     */
    next_key?: string;
    /**
     * total is total number of results available if PageRequest.count_total
     * was set, its value is undefined otherwise
     * @format uint64
     */
    total?: string;
  };
}

export interface GridParamsData {
  params?: {
    /** @format int64 */
    max_routes?: number;
  };
}

export interface RouteData {
  route?: {
    source?: string;
    destination?: string;
    name?: string;
    value?: {
      denom?: string;
      amount?: string;
    }[];
  };
}

export interface RoutesData {
  routes?: {
    source?: string;
    destination?: string;
    name?: string;
    value?: {
      denom?: string;
      amount?: string;
    }[];
  }[];
  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: {
    /**
     * next_key is the key to be passed to PageRequest.key to
     * query the next page most efficiently
     * @format byte
     */
    next_key?: string;
    /**
     * total is total number of results available if PageRequest.count_total
     * was set, its value is undefined otherwise
     * @format uint64
     */
    total?: string;
  };
}

export interface SourceRoutedEnergyData {
  value?: {
    denom?: string;
    amount?: string;
  }[];
}

export interface SourceRoutesData {
  routes?: {
    source?: string;
    destination?: string;
    name?: string;
    value?: {
      denom?: string;
      amount?: string;
    }[];
  }[];
  /**
   * PageResponse is to be embedded in gRPC response messages where the
   * corresponding request message has used PageRequest.
   *
   *  message SomeResponse {
   *          repeated Bar results = 1;
   *          PageResponse page = 2;
   *  }
   */
  pagination?: {
    /**
     * next_key is the key to be passed to PageRequest.key to
     * query the next page most efficiently
     * @format byte
     */
    next_key?: string;
    /**
     * total is total number of results available if PageRequest.count_total
     * was set, its value is undefined otherwise
     * @format uint64
     */
    total?: string;
  };
}

export interface InvestmintData {
  /**
   * Coin defines a token with a denomination and an amount.
   *
   * NOTE: The amount field is an Int which implements the custom method
   * signatures required by gogoproto.
   */
  amount?: {
    denom?: string;
    amount?: string;
  };
}

export interface ResourcesParamsData {
  params?: {
    /** @format int64 */
    max_slots?: number;
    /** @format int64 */
    halving_period_volt_blocks?: number;
    /** @format int64 */
    halving_period_ampere_blocks?: number;
    /** @format int64 */
    base_investmint_period_volt?: number;
    /** @format int64 */
    base_investmint_period_ampere?: number;
    /** @format int64 */
    min_investmint_period?: number;
    /**
     * Coin defines a token with a denomination and an amount.
     *
     * NOTE: The amount field is an Int which implements the custom method
     * signatures required by gogoproto.
     */
    base_investmint_amount_volt?: {
      denom?: string;
      amount?: string;
    };
    /**
     * Coin defines a token with a denomination and an amount.
     *
     * NOTE: The amount field is an Int which implements the custom method
     * signatures required by gogoproto.
     */
    base_investmint_amount_ampere?: {
      denom?: string;
      amount?: string;
    };
  };
}

export interface DmnParamsData {
  params?: {
    /** @format int64 */
    max_slots?: number;
    /** @format int64 */
    max_gas?: number;
    /** @format int64 */
    fee_ttl?: number;
  };
}

export interface ThoughtData {
  thought?: {
    program?: string;
    trigger?: {
      /** @format uint64 */
      period?: string;
      /** @format uint64 */
      block?: string;
    };
    load?: {
      input?: string;
      /**
       * Coin defines a token with a denomination and an amount.
       *
       * NOTE: The amount field is an Int which implements the custom method
       * signatures required by gogoproto.
       */
      gas_price?: {
        denom?: string;
        amount?: string;
      };
    };
    name?: string;
    particle?: string;
  };
}

export interface ThoughtStatsData {
  thought_stats?: {
    program?: string;
    name?: string;
    /** @format uint64 */
    calls?: string;
    /** @format uint64 */
    fees?: string;
    /** @format uint64 */
    gas?: string;
    /** @format uint64 */
    last_block?: string;
  };
}

export interface ThoughtsData {
  thoughts?: {
    program?: string;
    trigger?: {
      /** @format uint64 */
      period?: string;
      /** @format uint64 */
      block?: string;
    };
    load?: {
      input?: string;
      /**
       * Coin defines a token with a denomination and an amount.
       *
       * NOTE: The amount field is an Int which implements the custom method
       * signatures required by gogoproto.
       */
      gas_price?: {
        denom?: string;
        amount?: string;
      };
    };
    name?: string;
    particle?: string;
  }[];
}

export interface ThoughtsStatsData {
  thoughts_stats?: {
    program?: string;
    name?: string;
    /** @format uint64 */
    calls?: string;
    /** @format uint64 */
    fees?: string;
    /** @format uint64 */
    gas?: string;
    /** @format uint64 */
    last_block?: string;
  }[];
}